requirements.txt
RPi.GPIO
i2c_lcd
smbus2
paho_mqtt
wiringpi